ina_flask/requirements.txt
SQLAlchemy
Flask
inaSpeechSegmenter==0.6.8
youtube_dl
Keras==2.4.3
tensorflow==2.2.0
pandas
gunicorn
flask-cors
pyyaml
protobuf==3.20.0
watchfiles
mysql-connector-python